feat: add chat_message_queue table and migrate workflow events into it
P1 of the queue unification epic (#21267, closes #21334). Creates the
thread-keyed chat_message_queue table with typed items and moves the
workflow automation queue onto it, behavior-preserving:

- enqueue writes only chat_message_queue (item_type=workflow_event);
  drain claims legacy zero_workflow_queue_events rows first until the
  legacy table is empty (rolling-deploy dual-read, one release cycle)
- serialization moves to a per-thread advisory lock; the legacy
  per-workflow lock is still taken alongside while old API versions
  overlap
- pause state moves to chat_threads (queue_paused_at / pause_reason),
  dual-written to workflow_user_trigger_threads for old readers;
  workflow_user_trigger_threads degrades to a pure mapping table
- migration 0583 atomically moves pending events (id/created_at
  preserved) and copies pause state onto chat_threads
- workflow queue API/realtime response shapes unchanged
