+++ turbo/packages/db/src/index.ts
@@ -79,6 +79,7 @@ import * as runUploadedFileSchema from "./schema/run-uploaded-file";
 import * as vm0ApiKeySchema from "./schema/vm0-api-key";
 import * as zeroWorkflowSchema from "./schema/zero-workflow";
 import * as zeroWorkflowQueueSchema from "./schema/zero-workflow-queue";
+import * as chatMessageQueueSchema from "./schema/chat-message-queue";
 import * as computerUseHostSchema from "./schema/computer-use-host";
 import * as insightsDailySchema from "./schema/insights-daily";
 import * as userFeatureSwitchesSchema from "./schema/user-feature-switches";
@@ -187,6 +188,7 @@ export const schema = {
   ...vm0ApiKeySchema,
   ...zeroWorkflowSchema,
   ...zeroWorkflowQueueSchema,
+  ...chatMessageQueueSchema,
   ...computerUseHostSchema,
   ...insightsDailySchema,
   ...userFeatureSwitchesSchema,
